Patient Dashboard.jsx code added.
