--- client/src/pages/Dashboard/Dashboard.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Dashboard = () => {
-  return (
-    <div>
-      <h1 className="text-center">Dashbaord</h1>
-    </div>
-  );
-};
-
-export default Dashboard;
